requirements.txt
Flask
scikit-learn
mysql-connector-python
